refactor: added error handling

+++ backend/src/services/clients/DeleteClientsService.ts
@@ -1,6 +1,15 @@
+import { AppError } from '../../errors/AppError'
 import prismaCLient from '../../prisma'
 class DeleteClientsService {
   async execute(cnpj_cpf: string) {
+    const clientAlreadExists = await prismaCLient.client.findUnique({
+      where: {
+        cnpj_cpf,
+      },
+    })
+    if (!clientAlreadExists) {
+      throw new AppError('Cliente inválido')
+    }
     await prismaCLient.client.delete({
       where: {
         cnpj_cpf,
